2009-07-02 15:48 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * include/hbcomp.h   * include/hbcompdf.h   * include/hbapifs.h   * include/hbexprop.h   * include/hbmacro.h   * include/hbapi.h   * include/hbapiitm.h   * include/hbdate.h   * source/pp/ppcore.c   * source/vm/itemapi.c   * source/vm/arrays.c   * source/vm/extend.c   * source/common/expropt1.c   * source/rtl/philes.c   * source/rtl/empty.c   * source/rtl/minmax.c   * source/rtl/dates.c   * source/rtl/filesys.c   * source/rdd/dbfnsx/dbfnsx1.c   * source/rdd/dbfcdx/dbfcdx1.c   * source/rdd/dbf1.c   * source/rdd/dbffpt/dbffpt1.c   * source/rdd/hbsix/sxdate.c   * source/compiler/hbmain.c   * source/compiler/complex.c   * source/compiler/harbour.yyc   * source/compiler/harbour.y   * contrib/hbct/files.c   * contrib/hbct/dattime2.c   * contrib/hbct/datetime.c   * contrib/xhb/filestat.c   * contrib/hbtip/utils.c   * examples/hboleold/win_ole.c     * LONG -> long for date/time related values       (julian, date, type, millisec)     ; This is an effort to clean Harbour sources from       Windows/legacy-Clipper basic types and replace       them with own or ANSI C ones.       In the above case, usage wasn't consistent,       LONG and long were mixed.
  * source/main/harbour.c
  * source/vm/fm.c
    * LONG -> long for some stat counters.

+++ harbour/source/common/expropt1.c
@@ -337,7 +337,7 @@ HB_EXPR_PTR hb_compExprNewLong( HB_LONG lValue, HB_COMP_DECL )
    return pExpr;
 }
 
-HB_EXPR_PTR hb_compExprNewDate( LONG lDate, HB_COMP_DECL )
+HB_EXPR_PTR hb_compExprNewDate( long lDate, HB_COMP_DECL )
 {
    HB_EXPR_PTR pExpr;
 
@@ -345,14 +345,14 @@ HB_EXPR_PTR hb_compExprNewDate( LONG lDate, HB_COMP_DECL )
 
    pExpr = HB_COMP_EXPR_NEW( HB_ET_DATE );
 
-   pExpr->value.asDate.lDate = ( LONG ) lDate;
+   pExpr->value.asDate.lDate = lDate;
    pExpr->value.asDate.lTime = 0;
    pExpr->ValType = HB_EV_DATE;
 
    return pExpr;
 }
 
-HB_EXPR_PTR hb_compExprNewTimeStamp( LONG lDate, LONG lTime, HB_COMP_DECL )
+HB_EXPR_PTR hb_compExprNewTimeStamp( long lDate, long lTime, HB_COMP_DECL )
 {
    HB_EXPR_PTR pExpr;
 
@@ -661,7 +661,7 @@ HB_EXPR_PTR hb_compExprNewAliasExpr( HB_EXPR_PTR pAlias, HB_EXPR_PTR pExpList,
    pExpr->value.asAlias.pAlias    = pAlias;
    pExpr->value.asAlias.pExpList  = pExpList;
    pExpr->value.asAlias.pVar      = NULL;
-   
+
    if( pAlias->ExprType == HB_ET_MACRO )
    {
       /* Is it a special case &variable->( expressionList ) */
@@ -1421,7 +1421,7 @@ HB_EXPR_PTR hb_compExprSetGetBlock( HB_EXPR_PTR pExpr, HB_COMP_DECL )
    HB_EXPR_PTR pIIF;
    HB_EXPR_PTR pSet;
 
-   /* create {|var|  expression 
+   /* create {|var|  expression
     * NOTE: this is not a valid variable name so there will be no collisions
    */
    /* create var==NIL */
